Fix EditMatchInfo fragment and refactoring

+++ app/src/main/java/hr/vspr/dpasic/tenniswithme/activity/register_mvp/RegisterPresenterImpl.java
@@ -57,7 +57,9 @@ public class RegisterPresenterImpl implements RegisterPresenter {
                 if (response.isSuccessful()) {
                     //delete old access token from LocalDB
                     AccessToken tokenToDelete = SQLite.select().from(AccessToken.class).querySingle();
-                    tokenToDelete.delete();
+                    if (tokenToDelete != null) {
+                        tokenToDelete.delete();
+                    }
 
                     //save access token to LocalDB
                     AccessToken accessToken = response.body();
@@ -65,7 +67,9 @@ public class RegisterPresenterImpl implements RegisterPresenter {
 
                     //delete old credentials from LocalDB
                     Credentials credentialsToDelete = SQLite.select().from(Credentials.class).querySingle();
-                    credentialsToDelete.delete();
+                    if (credentialsToDelete != null) {
+                        credentialsToDelete.delete();
+                    }
 
                     //save credentials to LocalDB
                     Credentials credentials = new Credentials(registerViewModel.getEmail(), registerViewModel.getPassword());
